Wrap task hub deletion with activity.
